[BugFix] Fix some gradle UT failure
