+++ fe/fe-core/src/test/java/com/starrocks/planner/HiveTableSinkTest.java
@@ -26,6 +26,7 @@ import com.starrocks.connector.ConnectorMgr;
 import com.starrocks.connector.ConnectorTableId;
 import com.starrocks.connector.exception.StarRocksConnectorException;
 import com.starrocks.connector.hive.HiveStorageFormat;
+import com.starrocks.credential.CloudConfiguration;
 import com.starrocks.credential.CloudConfigurationFactory;
 import com.starrocks.qe.SessionVariable;
 import com.starrocks.sql.analyzer.AnalyzeTestUtil;
@@ -78,8 +79,10 @@ public class HiveTableSinkTest {
 
         new Expectations() {
             {
+                CloudConfiguration cloudConfig = new CloudConfiguration();
+                cloudConfig.loadCommonFields(new HashMap<>());
                 hiveConnector.getMetadata().getCloudConfiguration();
-                result = CloudConfigurationFactory.buildCloudConfigurationForStorage(new HashMap<>());
+                result = cloudConfig;
                 minTimes = 1;
             }
         };
