+++ fe/fe-core/src/test/java/com/starrocks/utframe/MockedBackend.java
@@ -167,12 +167,14 @@ public class MockedBackend {
     private static final AtomicInteger BASE_PORT = new AtomicInteger(8000);
     private static final long PATH_HASH = 123456;
 
+    private final int backendId;
     private final String host;
     private final int brpcPort;
     private final int heartBeatPort;
     private final int beThriftPort;
     private final int httpPort;
     private final int starletPort;
+
 
     final MockHeatBeatClient heatBeatClient;
 
@@ -182,11 +184,12 @@ public class MockedBackend {
 
     private final MockLakeService lakeService;
 
-    public MockedBackend(String host) {
-        this(host, BASE_PORT.getAndIncrement());
+    public MockedBackend(int backendId, String host) {
+        this(backendId, host, BASE_PORT.getAndIncrement());
     }
 
-    public MockedBackend(String host, int beThriftPort) {
+    public MockedBackend(int backendId, String host, int beThriftPort) {
+        this.backendId = backendId;
         this.host = host;
         this.beThriftPort = beThriftPort;
 
@@ -225,6 +228,10 @@ public class MockedBackend {
                 return backendService;
             }
         };
+    }
+
+    public int getBackendId() {
+        return backendId;
     }
 
     public String getHost() {
@@ -315,7 +322,7 @@ public class MockedBackend {
                         e.printStackTrace();
                     }
                 }
-            }).start();
+            }, "mock-be-thrift-client-" + backend.getBackendId()).start();
         }
 
         @Override
